\name{CTCallR-package}
\alias{CTCallR-package}
\alias{CTCallR}
\docType{package}
\title{
\packageTitle{CTCallR}
}
\description{
\packageDescription{CTCallR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CTCallR}
\packageIndices{CTCallR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{CTCallR}

Maintainer: \packageMaintainer{CTCallR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
